third_party/WebKit/LayoutTests/fast/dom/DOMImplementation/createDocumentType-err-expected.txt

createDocument tests modeled after mozilla's testing

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.implementation.createDocumentType('foo') threw exception TypeError: Failed to execute 'createDocumentType' on 'DOMImplementation': 3 arguments required, but only 1 present..
PASS document.implementation.createDocumentType('foo', null) threw exception TypeError: Failed to execute 'createDocumentType' on 'DOMImplementation': 3 arguments required, but only 2 present..
PASS document.implementation.createDocumentType(undefined, undefined) threw exception TypeError: Failed to execute 'createDocumentType' on 'DOMImplementation': 3 arguments required, but only 2 present..
PASS document.implementation.createDocumentType(null, undefined) threw exception TypeError: Failed to execute 'createDocumentType' on 'DOMImplementation': 3 arguments required, but only 2 present..
PASS document.implementation.createDocumentType(undefined, null) threw exception TypeError: Failed to execute 'createDocumentType' on 'DOMImplementation': 3 arguments required, but only 2 present..
PASS document.implementation.createDocumentType(undefined, undefined, null) did not throw exception.
PASS document.implementation.createDocumentType(null, null) threw exception TypeError: Failed to execute 'createDocumentType' on 'DOMImplementation': 3 arguments required, but only 2 present..
PASS document.implementation.createDocumentType(null, '') threw exception TypeError: Failed to execute 'createDocumentType' on 'DOMImplementation': 3 arguments required, but only 2 present..
PASS document.implementation.createDocumentType('', null) threw exception TypeError: Failed to execute 'createDocumentType' on 'DOMImplementation': 3 arguments required, but only 2 present..
PASS document.implementation.createDocumentType('', '') threw exception TypeError: Failed to execute 'createDocumentType' on 'DOMImplementation': 3 arguments required, but only 2 present..
PASS document.implementation.createDocumentType('a:', null, null) threw exception NamespaceError: An attempt was made to create or change an object in a way which is incorrect with regard to namespaces..
PASS document.implementation.createDocumentType(':foo', null, null) threw exception NamespaceError: An attempt was made to create or change an object in a way which is incorrect with regard to namespaces..
PASS document.implementation.createDocumentType(':', null, null) threw exception NamespaceError: An attempt was made to create or change an object in a way which is incorrect with regard to namespaces..
PASS document.implementation.createDocumentType('foo', null, null) did not throw exception.
PASS document.implementation.createDocumentType('foo:bar', null, null) did not throw exception.
PASS document.implementation.createDocumentType('foo::bar', null, null) threw exception NamespaceError: An attempt was made to create or change an object in a way which is incorrect with regard to namespaces..
PASS document.implementation.createDocumentType('	:bar', null, null) threw exception InvalidCharacterError: The string contains invalid characters..
PASS document.implementation.createDocumentType('foo:	', null, null) threw exception InvalidCharacterError: The string contains invalid characters..
PASS document.implementation.createDocumentType('foo :bar', null, null) threw exception InvalidCharacterError: The string contains invalid characters..
PASS document.implementation.createDocumentType('foo: bar', null, null) threw exception InvalidCharacterError: The string contains invalid characters..
PASS document.implementation.createDocumentType('a:b:c', null, null) threw exception NamespaceError: An attempt was made to create or change an object in a way which is incorrect with regard to namespaces..
PASS successfullyParsed is true

TEST COMPLETE
